Revision 8: Move FeedbackOption cells to annotation-queues directory and improve styling
- Move FeedbackOptionCell.tsx and FeedbackOptionCommentCell.tsx to pages-shared/annotation-queues/
- Update imports to use relative paths for co-located components
- Apply text styling to Comments cell (comet-body-s-accented text-muted-slate)
- Better organization: cells specific to annotation queues now located with ScoresContent

Improves code organization and visual consistency
